requirements.txt
asgiref==3.8.1
Django~=5.1.1
django-cors-headers~=4.4.0
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2

pip~=22.0.2
wheel~=0.37.1
protobuf~=3.12.4
cryptography~=3.4.8
Pillow~=9.0.1
keyring~=23.5.0
psycopg2~=2.9.9
psycopg2-binary~=2.9.9
colorama~=0.4.4
PyYAML~=5.4.1
numpy~=1.26.4
requests~=2.25.1
setuptools~=59.6.0
future~=0.18.2
pytz~=2022.1